backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
matplotlib==3.9.2
rapidfuzz==3.9.7
anthropic==0.34.2
requests==2.32.3
python-multipart==0.0.12
